moved dataclasses to main file
